Harmonize 2010 Decennial Census, 2019 ACS, and 2020 Block Population Estimates

Creates estimates for race/ethnicity variables for blocks in 2020, based on the 2010 Census (download_2010_vars()) and 2019 ACS (download_acs_vars()) data.

harmonize_vars(block_d, census_d, acs_d)

Arguments

block_d
the output of est_2020()

census_d
the output of download_2010_vars() or similar

acs_d
the output of download_acs_vars() or similar

Value

A data frame of 2020 block-level estimates for the variables in census_d.

Details

The racial breakdown of total and voting-age population are proportionally adjusted within block groups to match the ACS block-group percentages. Then these percentages are applied to the newly estimated block populations. Finally, the block race populations are rescaled so that the block populations match the 2020 estimates. This is essentially a 1-round iterative proportional fitting (IPF) procedure.